* master: (266 commits)
  Give a reason why the test was skipped
  Use variables that already received ZEND_NUM_ARGS()
  Return bool from zlog_stream_set_msg_{prefix,suffix}
  Fix #52974: jewish.c: compile error under Windows with GBK charset
  Update NEWS and UPGRADING with info about syslog.filter
  Rename none syslog.filter to all and update ini description
  Add syslog.filter INI for filtering syslog messages
  Add UPGRADING notes for deprecations
  Deprecate pdo_odbc.db2_instance_name
  Deprecate string.strip_tags filter
  Deprecate fgetss() and gzgetss()
  Deprecate defining a free-standing assert() function
  Deprecate non-string needles in string search functions
  Deprecate undocumented mbereg_* aliases
  Move GMP bug 74670 test file to tests dir
  Fix off-by-one allocation
  Fix incorrect length calculation
  [ci skip] Add new global MBString constants to UPGRADING
  Removed and fixed unused variables
  Update binary SDK version for AppVeyor
  ...

+++ ext/readline/readline_cli.c
@@ -208,15 +208,15 @@ static zend_string *cli_get_prompt(char *block, char prompt) /* {{{ */
 }
 /* }}} */
 
-static int cli_is_valid_code(char *code, int len, zend_string **prompt) /* {{{ */
+static int cli_is_valid_code(char *code, size_t len, zend_string **prompt) /* {{{ */
 {
 	int valid_end = 1, last_valid_end;
 	int brackets_count = 0;
 	int brace_count = 0;
-	int i;
+	size_t i;
 	php_code_type code_type = body;
 	char *heredoc_tag;
-	int heredoc_len;
+	size_t heredoc_len;
 
 	for (i = 0; i < len; ++i) {
 		switch(code_type) {
@@ -403,7 +403,7 @@ static int cli_is_valid_code(char *code, int len, zend_string **prompt) /* {{{ *
 }
 /* }}} */
 
-static char *cli_completion_generator_ht(const char *text, int textlen, int *state, HashTable *ht, void **pData) /* {{{ */
+static char *cli_completion_generator_ht(const char *text, size_t textlen, int *state, HashTable *ht, void **pData) /* {{{ */
 {
 	zend_string *name;
 	zend_ulong number;
@@ -429,7 +429,7 @@ static char *cli_completion_generator_ht(const char *text, int textlen, int *sta
 	return NULL;
 } /* }}} */
 
-static char *cli_completion_generator_var(const char *text, int textlen, int *state) /* {{{ */
+static char *cli_completion_generator_var(const char *text, size_t textlen, int *state) /* {{{ */
 {
 	char *retval, *tmp;
 	zend_array *symbol_table = &EG(symbol_table);
@@ -444,7 +444,7 @@ static char *cli_completion_generator_var(const char *text, int textlen, int *st
 	return retval;
 } /* }}} */
 
-static char *cli_completion_generator_ini(const char *text, int textlen, int *state) /* {{{ */
+static char *cli_completion_generator_ini(const char *text, size_t textlen, int *state) /* {{{ */
 {
 	char *retval, *tmp;
 
@@ -458,7 +458,7 @@ static char *cli_completion_generator_ini(const char *text, int textlen, int *st
 	return retval;
 } /* }}} */
 
-static char *cli_completion_generator_func(const char *text, int textlen, int *state, HashTable *ht) /* {{{ */
+static char *cli_completion_generator_func(const char *text, size_t textlen, int *state, HashTable *ht) /* {{{ */
 {
 	zend_function *func;
 	char *retval = cli_completion_generator_ht(text, textlen, state, ht, (void**)&func);
@@ -470,7 +470,7 @@ static char *cli_completion_generator_func(const char *text, int textlen, int *s
 	return retval;
 } /* }}} */
 
-static char *cli_completion_generator_class(const char *text, int textlen, int *state) /* {{{ */
+static char *cli_completion_generator_class(const char *text, size_t textlen, int *state) /* {{{ */
 {
 	zend_class_entry *ce;
 	char *retval = cli_completion_generator_ht(text, textlen, state, EG(class_table), (void**)&ce);
@@ -482,7 +482,7 @@ static char *cli_completion_generator_class(const char *text, int textlen, int *
 	return retval;
 } /* }}} */
 
-static char *cli_completion_generator_define(const char *text, int textlen, int *state, HashTable *ht) /* {{{ */
+static char *cli_completion_generator_define(const char *text, size_t textlen, int *state, HashTable *ht) /* {{{ */
 {
 	zend_class_entry **pce;
 	char *retval = cli_completion_generator_ht(text, textlen, state, ht, (void**)&pce);
@@ -508,7 +508,7 @@ TODO:
 - future: respect scope ("php > function foo() { $[tab]" should only expand to local variables...)
 */
 	char *retval = NULL;
-	int textlen = strlen(text);
+	size_t textlen = strlen(text);
 
 	if (!index) {
 		cli_completion_state = 0;
@@ -519,7 +519,7 @@ TODO:
 		retval = cli_completion_generator_ini(text, textlen, &cli_completion_state);
 	} else {
 		char *lc_text, *class_name_end;
-		int class_name_len;
+		size_t class_name_len;
 		zend_string *class_name;
 		zend_class_entry *ce = NULL;
 
@@ -563,7 +563,7 @@ TODO:
 			zend_string_release_ex(class_name, 0);
 		}
 		if (ce && retval) {
-			int len = class_name_len + 2 + strlen(retval) + 1;
+			size_t len = class_name_len + 2 + strlen(retval) + 1;
 			char *tmp = malloc(len);
 
 			snprintf(tmp, len, "%s::%s", ZSTR_VAL(ce->name), retval);
